+++ src/android.rs
@@ -5,18 +5,18 @@ use core::ptr;
 ///
 /// ## Construction
 /// ```
-/// # use raw_window_handle::AndroidNDKHandle;
-/// let mut handle = AndroidNDKHandle::empty();
+/// # use raw_window_handle::AndroidNdkHandle;
+/// let mut handle = AndroidNdkHandle::empty();
 /// /* set fields */
 /// ```
 #[non_exhaustive]
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub struct AndroidNDKHandle {
+pub struct AndroidNdkHandle {
     /// A pointer to an `ANativeWindow`.
     pub a_native_window: *mut c_void,
 }
 
-impl AndroidNDKHandle {
+impl AndroidNdkHandle {
     pub fn empty() -> Self {
         Self {
             a_native_window: ptr::null_mut(),